18.03.24

+++ AllBumbers.cpp
@@ -9,7 +9,7 @@ using namespace std;
 
 int main()
 {
-    int n, a = 0, b = 0, i = 1;
+    long long n, a = 0, b = 0, i = 1;
     cin >> n;
 
     while (i <= n) {
